# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "sChat REST server",
        "title": "sChat Server API",
        "contact": {
            "name": "mykola.sainyuk@gmail.com",
            "url": "https://www.linkedin.com/in/mykola-sainiuk-3b03168b/"
        },
        "version": "1.0"
    },
    "host": "localhost:5000",
    "basePath": "/",
    "paths": {
...
        "/api/user/contact/add": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Adding another user as a contact",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "contact"
                ],
                "summary": "Add contact to User",
                "parameters": [
                    {
                        "description": "New contact input",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AddContactInputDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not found user",
                        "schema": {
                            "$ref": "#/definitions/httpexp.HttpExp"
                        }
                    }
                }
            }
        },
        "/api/user/contact/list": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Paginated list of User contacts",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "contact"
                ],
                "summary": "List contacts paginated",
                "parameters": [
                    {
                        "type": "string",
                        "description": "page number",
                        "name": "page",
                        "in": "path"
                    },
                    {
                        "type": "string",
                        "description": "page size",
                        "name": "limit",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.UserInfoOutputDto"
                            }
                        }
                    },
                    "404": {
                        "description": "Not found user",
                        "schema": {
                            "$ref": "#/definitions/httpexp.HttpExp"
                        }
                    }
                }
            }
        },
        "/api/user/contact/list/all": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Unpaginated list of all contacts of User",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "contact"
                ],
                "summary": "List all contacts",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.UserInfoOutputDto"
                            }
                        }
                    },
                    "404": {
                        "description": "Not found user",
                        "schema": {
                            "$ref": "#/definitions/httpexp.HttpExp"
                        }
                    }
                }
            }
        },
        "/api/user/me": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "ME endpoint",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Get user info",
                "responses": {
                    "200": {
                        "description": "User object extended",
                        "schema": {
                            "$ref": "#/definitions/dto.UserInfoExtendedOutputDto"
                        }
                    },
                    "404": {
                        "description": "Not found user",
                        "schema": {
                            "$ref": "#/definitions/httpexp.HttpExp"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.AddContactInputDto": {
            "type": "object",
            "required": [
                "username"
            ],
            "properties": {
                "username": {
                    "type": "string",
                    "minLength": 2
                }
            }
        },
        "dto.ChatOutputDto": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "iconUri": {
                    "type": "string"
                },
                "lastMessage": {
                    "type": "string"
                },
                "muted": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "users": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "dto.UserInfoExtendedOutputDto": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string"
                },
                "avatarUri": {
                    "type": "string"
                },
                "chats": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.ChatOutputDto"
                    }
                },
                "contacts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.UserInfoExtendedOutputDto"
                    }
                },
                "createdAt": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "dto.UserInfoOutputDto": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string"
                },
                "avatarUri": {
                    "type": "string"
                },
                "chats": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "contacts": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "createdAt": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "httpexp.HttpExp": {
            "description": "common 422 response",
            "type": "object",
            "properties": {
                "details": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}